Update job configs / exclude template from ID
If an existing job is found, it should be allowed to update.
Without this, the Jenkins job configuration dashboard grows out of
hand with new jobs any time job templates change.

As such, MD5 job id should not include the job file contents.

There is enough uniqueness in job name generation without resorting
to reading job file contents to affect the MD5 id.

This is pretty problematic if minor tweaks are made to the Jenkins
job.xml.erb templates.  Consider changes such as modifying email
notifications, or enabling Chuck Norris, etc.

We refresh jobs automatically as part of an automated deployment,
but this will result in new MD5 job ids created in Jenkins.  This
is bad news, especially when there are many jobs.  All jobs are
now effectively duplicated and the history is now no longer
correlated between the runs, etc.