Truncate MD5 digests in job names to 12 characters
On Windows, file paths are limited to 260 characters (including the
drive letter and null terminator). Projects get checked out to
directories with names like:

    C:\<jenkins root>\workspace\<job name>

Including a 32-character MD5 digest (plus the repo owner and name) in
the job name significantly reduces the number of characters left for use
by the files in the repository. (For example, for
brightray/libchromiumcontent the job name is 61 characters long. When
combined with the drive letter, null terminator, and `\workspace\`, only
184 characters are left for paths within the repository itself.)

So now we truncate MD5 digests in job names to just the 12 most
significant hex digits. This is enough for building
brightray/libchromiumcontent (for now at least).