Update the test runner to work with newer rubies
Ruby versions after 2.1.x remove the `testrb` script that used to shift
with Ruby by default. Due to its removal, copy the contents of the
script from an older version of Ruby and use it in our `script/test`
script so that we can upgrade Ruby without having to switch out the test
framework and setup.